app/pyproject.toml
[project]
name = "test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    { include-group = "test" },
]
test = [
    "factory-boy>=3.3.3",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.3.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
]

[tool.pytest]
testpaths = [
    "tests",
]
markers = [
    "integration: integration tests",
    "unit: unit tests"
]
addopts = [
    "-s",
    "-v",
    "--color=yes",
    "--cov=src",
    "--cov=tests",
    "--cov-report=html:coverage/htmlcov",
    "--cov-report=xml:coverage/coverage.xml",
    "--cov-report=term",
    "--cov-fail-under=100",
    "--cov-context=test", ]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "session"

[tool.coverage.run]
data_file = "coverage/.coverage"
parallel = true
branch = true
concurrency = ["greenlet", "thread", "multiprocessing"]

[tool.coverage.html]
show_contexts = true
